Split out breadcrumbs class
